Refactor TaskList component

+++ src/components/TaskMenu.jsx
@@ -1,11 +1,11 @@
-import React, { useRef, useState } from "react"
+import { forwardRef, useRef, useState } from "react"
 import * as Dialog from "@radix-ui/react-dialog"
 import * as Menu from "@radix-ui/react-dropdown-menu"
 
 const API_BASE = import.meta.env.VITE_API_BASE
 
-const TaskMenu = ({ task }) => {
-  const [menuOpen, setMenuOpen] = useState(false)
+const TaskMenu = ({ task, setTask }) => {
+  const [open, setOpen] = useState(false)
   const [hasOpenDialog, setHasOpenDialog] = useState(false)
   const menuTriggerRef = useRef(null)
   const focusRef = useRef(null)
@@ -17,22 +17,24 @@ const TaskMenu = ({ task }) => {
   const handleDialogItemOpenChange = (open) => {
     setHasOpenDialog(open)
     if (open === false) {
-      setMenuOpen(false)
+      setOpen(false)
     }
   }
 
   return (
-    <Menu.Root open={menuOpen} onOpenChange={setMenuOpen}>
+    <Menu.Root open={open} onOpenChange={setOpen}>
       <Menu.Trigger asChild>
         <button
           ref={menuTriggerRef}
-          className="w-8 h-8 block px-2.5 py-1.5 rounded-sm border border-gray-200 hover:bg-gray-200 dark:border-gray-800 dark:hover:bg-gray-800"
+          className="w-8 h-8 grid place-items-center rounded-sm border border-gray-200 hover:bg-gray-200 dark:border-gray-800 dark:hover:bg-gray-800"
         >
           <div className="-translate-y-1 text-gray-500">...</div>
         </button>
       </Menu.Trigger>
+
       <Menu.Content
         hidden={hasOpenDialog}
+        className="bg-white dark:bg-gray-900 rounded-md shadow shadow-menu p-2 translate-y-1 z-10 animate-menu-content"
         onCloseAutoFocus={(event) => {
           if (focusRef.current) {
             focusRef.current.focus()
@@ -40,16 +42,17 @@ const TaskMenu = ({ task }) => {
             event.preventDefault()
           }
         }}
-        className="bg-white dark:bg-gray-900 rounded-md shadow shadow-menu p-2 translate-y-1 z-10 animate-menu-content"
       >
         <div className="flex flex-col gap-2">
           <TaskEditDialog
             task={task}
+            setTask={setTask}
             onSelect={handleDialogItemSelect}
             onOpenChange={handleDialogItemOpenChange}
           />
           <TaskDeleteDialog
             task={task}
+            setTask={setTask}
             onSelect={handleDialogItemSelect}
             onOpenChange={handleDialogItemOpenChange}
           />
@@ -60,75 +63,50 @@ const TaskMenu = ({ task }) => {
   )
 }
 
-const DialogItem = React.forwardRef((props, forwardedRef) => {
-  const { triggerChildren, children, onSelect, onOpenChange, ...itemProps } =
-    props
-  return (
-    <Dialog.Root onOpenChange={onOpenChange}>
-      <Dialog.Trigger asChild>
-        <Menu.Item
-          {...itemProps}
-          ref={forwardedRef}
-          className="MenuItem"
-          onSelect={(event) => {
-            event.preventDefault()
-            onSelect && onSelect()
-          }}
-        >
-          {triggerChildren}
-        </Menu.Item>
-      </Dialog.Trigger>
-      <Dialog.Portal>
-        <Dialog.Overlay className="bg-black bg-opacity-20 fixed inset-0" />
-        <Dialog.Content className="bg-white dark:bg-gray-900 dark:text-white fixed top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 max-w-sm w-11/12 rounded-lg shadow-xl p-6 animate-dialog-content">
-          {children}
-        </Dialog.Content>
-      </Dialog.Portal>
-    </Dialog.Root>
-  )
-})
-
-DialogItem.displayName = "DialogItem"
-
-const TaskEditDialog = ({ task, onSelect, onOpenChange }) => {
+const TaskEditDialog = ({ task, setTask, onSelect, onOpenChange }) => {
   const [newName, setNewName] = useState(task.name)
 
-  const handleNewNameChange = (event) => setNewName(event.target.value)
+  const handleNameChange = (event) => {
+    setNewName(event.target.value)
+  }
 
   const handleSubmit = (event) => {
     event.preventDefault()
 
+    setTask({ ...task, name: newName })
+
     const token = localStorage.getItem("token")
-    const options = {}
-    options.method = "PUT"
-    options.headers = {
-      "Content-Type": "application/json",
-      Authorization: `Bearer ${token}`,
+    if (token) {
+      const options = {}
+      options.method = "PUT"
+      options.headers = {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
+      }
+      options.body = JSON.stringify({ name: newName })
+      fetch(`${API_BASE}/tasks/${task._id}`, options).catch((err) =>
+        console.error(err)
+      )
     }
-    options.body = JSON.stringify({ name: newName })
-    fetch(`${API_BASE}/tasks/${task._id}`, options)
-      .then(() => location.reload())
-      .catch((err) => console.error(err))
+
+    onOpenChange(false)
   }
 
   return (
     <DialogItem
-      triggerChildren={
-        <button className="block w-full font-semibold text-sm text-gray-500 bg-gray-100 hover:bg-gray-200 px-6 py-2 rounded dark:bg-gray-800 dark:text-gray-400 dark:hover:bg-gray-700">
-          Edit
-        </button>
-      }
+      triggerChildren={<button className="btn small secondary">Edit</button>}
       onSelect={onSelect}
       onOpenChange={onOpenChange}
     >
       <Dialog.Title className="font-bold text-2xl mb-4">Edit task</Dialog.Title>
+
       <form onSubmit={handleSubmit}>
         <input
-          className="block w-full mb-4 rounded border-gray-300 bg-transparent dark:border-gray-600"
+          className="block w-full mb-6 rounded border-gray-300 bg-transparent dark:border-gray-600"
           type="text"
           name="name"
           value={newName}
-          onChange={handleNewNameChange}
+          onChange={handleNameChange}
         />
         <div className="flex justify-end gap-4">
           <Dialog.Close asChild>
@@ -139,7 +117,10 @@ const TaskEditDialog = ({ task, onSelect, onOpenChange }) => {
               Close
             </button>
           </Dialog.Close>
-          <button className="font-semibold text-white bg-blue-700 hover:bg-blue-800 px-6 py-2 rounded">
+          <button
+            type="submit"
+            className="font-semibold text-white bg-blue-700 hover:bg-blue-800 px-6 py-2 rounded"
+          >
             Save
           </button>
         </div>
@@ -148,16 +129,23 @@ const TaskEditDialog = ({ task, onSelect, onOpenChange }) => {
   )
 }
 
-const TaskDeleteDialog = ({ task, onSelect, onOpenChange }) => {
+const TaskDeleteDialog = ({ task, setTask, onSelect, onOpenChange }) => {
   const handleDelete = () => {
+    setTask(null)
+
     const token = localStorage.getItem("token")
-    const options = {}
-    options.method = "DELETE"
-    options.headers = { Authorization: `Bearer ${token}` }
-    fetch(`${API_BASE}/tasks/${task._id}`, options)
-      .then(() => location.reload())
-      .catch((err) => console.error(err))
+    if (token) {
+      const options = {}
+      options.method = "DELETE"
+      options.headers = { Authorization: `Bearer ${token}` }
+      fetch(`${API_BASE}/tasks/${task._id}`, options).catch((err) =>
+        console.error(err)
+      )
+    }
+
+    onOpenChange(false)
   }
+
   return (
     <DialogItem
       triggerChildren={
@@ -171,7 +159,7 @@ const TaskDeleteDialog = ({ task, onSelect, onOpenChange }) => {
       <Dialog.Title className="font-bold text-2xl mb-4">
         Delete task
       </Dialog.Title>
-      <p className="text-gray-600 mb-4">
+      <p className="text-gray-600 dark:text-gray-400 mb-6">
         Do you really want to delete that{" "}
         <span className="text-black dark:text-white">"{task.name}"</span> task?
       </p>
@@ -194,5 +182,37 @@ const TaskDeleteDialog = ({ task, onSelect, onOpenChange }) => {
     </DialogItem>
   )
 }
+
+const DialogItem = forwardRef((props, forwardedRef) => {
+  const { triggerChildren, children, onSelect, onOpenChange, ...itemProps } =
+    props
+
+  return (
+    <Dialog.Root onOpenChange={onOpenChange}>
+      <Dialog.Trigger asChild>
+        <Menu.Item
+          {...itemProps}
+          ref={forwardedRef}
+          className="MenuItem"
+          onSelect={(event) => {
+            event.preventDefault()
+            onSelect && onSelect()
+          }}
+        >
+          {triggerChildren}
+        </Menu.Item>
+      </Dialog.Trigger>
+
+      <Dialog.Portal>
+        <Dialog.Overlay className="bg-black bg-opacity-20 fixed inset-0" />
+        <Dialog.Content className="bg-white dark:bg-gray-900 dark:text-white fixed top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 max-w-sm w-11/12 rounded-lg shadow-xl p-6 animate-dialog-content">
+          {children}
+        </Dialog.Content>
+      </Dialog.Portal>
+    </Dialog.Root>
+  )
+})
+
+DialogItem.displayName = "Dialog item"
 
 export default TaskMenu
